# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce Order & Promotion API",
        "description": "REST API for browsing products, managing a cart, applying promotional codes, checking out and managing orders.\n\n**Money** is returned as decimal strings (e.g. `\"149.99\"`) and accepted in dollars; it is stored as integer cents.\n\n**Errors** share one shape: `{ \"message\", \"code\", \"details\"? }`. Validation errors (`validation_failed`) add an `errors` object keyed by field.\n\n**Authentication**: call `POST /api/auth/login` (or register), then send `Authorization: Bearer <token>`.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "This server"
        }
    ],
    "paths": {
        "/api/admin/products": {
            "get": {
                "tags": [
                    "Admin: Products"
                ],
                "summary": "List all products, including inactive",
                "description": "All products, including inactive ones, newest first.",
                "operationId": "adminListProducts",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/PerPage"
                    },
                    {
                        "$ref": "#/components/parameters/Page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated products, newest first.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationFailed"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin: Products"
                ],
                "summary": "Create a product",
                "operationId": "adminCreateProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationFailed"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/products/{product}": {
            "get": {
                "tags": [
                    "Admin: Products"
                ],
                "summary": "Show any product",
                "operationId": "adminShowProduct",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product id.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The product.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin: Products"
                ],
                "summary": "Update a product",
                "description": "Also available as PUT. Send only the fields to change. Products are not deleted; set status to inactive instead.",
                "operationId": "adminUpdateProduct",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product id.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationFailed"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "required": [
                    "message",
                    "code"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Only 3 units of 'Desk Lamp' are available."
                    },
                    "code": {
                        "description": "Machine-readable error code.",
                        "type": "string",
                        "example": "insufficient_stock"
                    },
                    "details": {
                        "description": "Extra context for some errors.",
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": true
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "required": [
                    "message",
                    "code",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The quantity field must be at least 1."
                    },
                    "code": {
                        "type": "string",
                        "example": "validation_failed"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "quantity": [
                                "The quantity field must be at least 1."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 4
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    },
                    "total": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Walnut Desk Lamp"
                    },
                    "sku": {
                        "type": "string",
                        "example": "LAMP-001"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "string",
                        "example": "49.99"
                    },
                    "stock_quantity": {
                        "type": "integer",
                        "example": 5
                    },
                    "in_stock": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ProductInput": {
                "description": "All fields required on create (POST); any subset on update (PATCH/PUT).",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Walnut Desk Lamp",
                        "maxLength": 255
                    },
                    "sku": {
                        "description": "Unique; letters, numbers, dots, dashes, underscores.",
                        "type": "string",
                        "example": "LAMP-001",
                        "maxLength": 64
                    },
                    "description": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 5000
                    },
                    "price": {
                        "description": "Dollars, up to 2 decimals.",
                        "type": "string",
                        "example": "49.99"
                    },
                    "stock_quantity": {
                        "type": "integer",
                        "example": 5,
                        "minimum": 0
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "Unauthenticated": {
                "description": "Missing or invalid bearer token (`unauthenticated`).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "example": {
                            "message": "Unauthenticated.",
                            "code": "unauthenticated"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Authenticated but not allowed (`forbidden`), e.g. another customer's order or a non-admin on an admin route.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "example": {
                            "message": "You do not have access to this order.",
                            "code": "forbidden"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "The resource does not exist (`product_not_found`, `order_not_found`, `cart_item_not_found`, `promotion_not_found`, `not_found`).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "example": {
                            "message": "Product not found.",
                            "code": "product_not_found"
                        }
                    }
                }
            },
            "ValidationFailed": {
                "description": "The input is invalid (`validation_failed`).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "TooManyRequests": {
                "description": "Rate limit exceeded (`too_many_requests`); see the `Retry-After` header.",
                "headers": {
                    "Retry-After": {
                        "description": "Seconds until another request is allowed.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                },
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "example": {
                            "message": "Too Many Attempts.",
                            "code": "too_many_requests"
                        }
                    }
                }
            }
        },
        "parameters": {
            "PerPage": {
                "name": "per_page",
                "in": "query",
                "description": "Items per page (1–100).",
                "schema": {
                    "type": "integer",
                    "default": 15,
                    "maximum": 100,
                    "minimum": 1
                }
            },
            "Page": {
                "name": "page",
                "in": "query",
                "description": "Page number.",
                "schema": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum personal access token from /api/auth/login or /api/auth/register.",
                "bearerFormat": "Sanctum token",
                "scheme": "bearer"
            }
        }
    },
...
}